restructuring for separate lambdas
