+++ src/functions/receive/index.js
@@ -3,8 +3,8 @@ const crypto = require('crypto');
 require('dotenv').config();
 
 // Function for sending twitter DMs
-const twitter = require('./lib/twitter/');
-import cards from './lib/cards/';
+import twitter from '../../lib/twitter/'
+import cards from '../../lib/cards/';
 
 const twitterUserId = '905602080252977152' //@riskmapus bot
 
@@ -139,32 +139,3 @@ module.exports.twitterDMWebhook = (event, context, callback) => {
     }
   }
 }
-
-module.exports.twitterReply = (event, context, callback) => {
-  //This module listens in to SNS Twitter topic and reads the message published
-  var message = JSON.parse(event.Records[0].Sns.Message);
-  console.log('Message received from SNS topic: ' + message);
-  // Prepare message
-  let msg = {};
-  msg.event = {
-        "type": "message_create",
-        "message_create": {
-          "target": {
-            "recipient_id": message.username
-          },
-          "message_data": {
-            "text": "Thank you for your report. You can access it using this link https://riskmap.us/map/broward/" + message.report_id
-          }
-        }
-      }
-
-  //Construct the confirmation message to be sent to the user
-  //var messageText = getConfirmationMessageText(message.language, message.implementation_area, message.report_id);
-  //var messageText = '@' + message.username + ' ' + messageText;
-
-  //Make a POST call to send a tweet to the user
-  //sendTweet(messageText, message.username);
-  twitter.sendMessage(msg, function(err, response){
-
-  })
-};
