+++ src/lib/cards/index.js
@@ -12,8 +12,6 @@ import request from 'request'
  * @param {Object} body Request body object or null
  **/
 const processResponse = (error, response, body) => new Promise ((resolve, reject) =>{
-  console.log('err: ' + JSON.stringify(error));
-  console.log('res: ' + JSON.stringify(response));
   if (!error && response.statusCode === 200){
     resolve(body.cardId)
   } else {
@@ -33,7 +31,6 @@ export default () => ({
   * @param {String} language User's language as two letter code (e.g. 'en')
   **/
   getCardLink: (username, network, language) => new Promise((resolve, reject) => {
-    console.log('get card link function fired')
     let cardRequest = {
       "username": username,
       "network": network,
@@ -51,42 +48,10 @@ export default () => ({
       json: true,
       body: cardRequest
     }, function(error, response, body) {
-      //return processResponse(error, response, body);
-      console.log('err 1: ' + JSON.stringify(error));
-      console.log('res 1: ' + JSON.stringify(response));
+      // Process response
       processResponse(error, response, body)
         .then((cardId) => resolve(cardId))
         .catch((err) => reject(err))
     });
   })
 });
-
-/*let getCardLink = function(username, network, language, callback) {
-  var card_request = {
-    "username": username,
-    "network": network,
-    "language": language
-  };
-
-  // Get a card from Cognicity server
-  request({
-    url: 'https://3m3l15fwsf.execute-api.us-west-2.amazonaws.com/prod/cards',
-    method: 'post',
-    headers: {
-      'content-type': 'application/json',
-      'x-api-key': process.env.SERVER_API_KEY
-    },
-    port: 443,
-    json: true,
-    body: card_request
-  }, function(error, response, body) {
-    if (!error && response.statusCode === 200) {
-      callback(null, body.cardId); //Return cardId on success
-    } else {
-      var err = 'Error getting card: ' + JSON.stringify(error) + JSON.stringify(response);
-      callback(err, null); // Return error
-    }
-  });
-}*/
-
-//module.exports = {getCardLink};
